Added FH5 ready check
slash command to send a embed with a single button with a hardcoded role for ready check to confirm entrants participation before the event.
Loops trough the members of the server to remove role from all that may have it.
can be run from anywhere as the target send channel is also coded in.